# swagger_local.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ecommerce Api server for admin and web",
    "version": "1.0.0",
    "title": "Ecommerce"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/auth/login/user": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member login check",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "member username and password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/member/register": {
      "post": {
        "tags": [
          "admin member"
        ],
        "summary": "admin",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "member register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/member/user": {
      "get": {
        "tags": [
          "admin user"
        ],
        "summary": "admin",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/member/user/{userId}": {
      "get": {
        "tags": [
          "admin user"
        ],
        "summary": "admin",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "admin user"
        ],
        "summary": "update user by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "admin user"
        ],
        "summary": "Deletes an admin",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
    "/api/v1/member": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "active user data",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/member/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "register user",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "member register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/verify": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Member verify",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/member/update/{userId}": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "update user",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/member/resetPassword": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Member Password reset",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateUser": {
      "type": "object",
      "properties": {
        "fname": {
          "type": "string",
          "example": "himal"
        },
        "lname": {
          "type": "string",
          "example": "shrestha"
        },
        "city": {
          "type": "string",
          "example": "circle"
        },
        "state": {
          "type": "string",
          "example": "somwhere wlse"
        },
        "zip": {
          "type": "string",
          "example": "8945"
        },
        "phone": {
          "type": "string",
          "example": "1223234234"
        },
        "country": {
          "type": "string",
          "example": "nepal"
        },
        "address": {
          "type": "string",
          "example": "far"
        },
        "address2": {
          "type": "string",
          "example": "somewhere"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        },
        "role": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "UpdateUser"
      }
    },
    "RegisterAdmin": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himalhs1@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "himal123"
        },
        "email": {
          "type": "string",
          "example": "himal123"
        },
        "roles": {
          "type": "string",
          "example": "himal123"
        },
        "fname": {
          "type": "string",
          "example": "himal123"
        },
        "lname": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "RegisterAdmin"
      }
    },
    "RegisterUser": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himalhs1@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "himal123"
        },
        "fname": {
          "type": "string",
          "example": "himal123"
        },
        "lname": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "RegisterUser"
      }
    },
    "Auth": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himal"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "Pet"
      }
    }
  },
...
}